+++ src/GUI/MainWindow.cpp
@@ -45,7 +45,7 @@ bool getFreeDiskSpace(QString path, QString &result){
 }
 
 MainWindow::MainWindow(const QMap<QString, QString> &urlArgs, const QString &dbPath, QWidget *parent) :
-        PXMainWindow("Software", QIcon::fromTheme("panther"), parent){
+        PXMainWindow("Software", QIcon::fromTheme("panther"), "px-desktop-wiki:page=PantherX-Software/index.html", parent){
     searchBox()->addAction(QIcon::fromTheme("system-search"),QLineEdit::TrailingPosition);
     CacheManager::init(CACHE_DIR);
     CacheManager::instance()->clear();
@@ -243,10 +243,6 @@ void MainWindow::searchBoxTextEditedHandler(PXContentWidget *currentWidget, cons
 
 void MainWindow::settingsButtonHandler() {
     QDesktopServices::openUrl(QUrl("px-settings-ui:module=software"));
-}
-
-void MainWindow::helpButtonHandler(){
-    qDebug() << "TODO - Help Button Handler";
 }
 
 void MainWindow::showTerminalSignalHandler(TerminalWidget *terminal){