+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 ### Fixed and Changed
  
  - DB path relocation.
- 
+ - Update Other Applications layout.
 
 ## [0.1.8]
 ### Fixed and Changed